requirements.txt
torch>=1.7
torchvision>=0.5.0
inplace-abn
pycocotools
randaugment~=1.0.2

matplotlib~=3.4.3
numpy~=1.21.3
pillow~=8.4.0
mmcv~=1.6.1
timm~=0.4.12
termcolor~=1.1.0
pandas~=1.3.5
sklearn~=0.0
scikit-learn~=1.0.1